Journalist interface: Add NewUserForm
Validate form inputs using WTForms. This fixes #2500, where an odd
length input would throw a 500 error.

We also use a custom validator for OTP secret

Whitespace should be ignored when computing the 40 characters required for
the OTP secret.

We can use a custom validator:

https://wtforms.readthedocs.io/en/latest/validators.html#custom-validators

to ensure this field is correct.